Reorganize Calc tests to match module names
As requested by the user, the monolithic `test_calc_uno.py` has been split
into smaller files matching the modules in `plugin/calc/`. Tests that execute
pure logic (without needing the UNO context) have been stripped of the
`_uno` suffix and the `@native_test` decorator to make them discoverable by standard pytest.
The test `test_orca_catalog.py` has been correctly moved to `scripts/tests/`.
