+++ plugin/tests/calc/test_cells.py
@@ -1,0 +1,20 @@
+# WriterAgent - AI Writing Assistant for LibreOffice
+# Copyright (c) 2024 John Balis
+# Copyright (c) 2026 KeithCu (modifications and relicensing)
+# Copyright (c) 2026 LibreCalc AI Assistant (Calc integration features, originally MIT)
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+import unittest
+
+
+def test_cells_parse_color():
+    from plugin.calc.cells import _parse_color
+    assert _parse_color("red") == 0xFF0000
+    assert _parse_color("RED") == 0xFF0000
+    assert _parse_color("#00FF00") == 0x00FF00
+    assert _parse_color("#000") == 0x000000
+    assert _parse_color("invalid") is None
