+++ plugin/tests/calc/test_charts_uno.py
@@ -1,0 +1,109 @@
+# WriterAgent - AI Writing Assistant for LibreOffice
+# Copyright (c) 2024 John Balis
+# Copyright (c) 2026 KeithCu (modifications and relicensing)
+# Copyright (c) 2026 LibreCalc AI Assistant (Calc integration features, originally MIT)
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+from plugin.framework.uno_context import get_desktop
+from plugin.testing_runner import setup, teardown, native_test
+
+_test_doc = None
+_test_ctx = None
+
+@setup
+def setup_calc_tests(ctx):
+    global _test_doc, _test_ctx
+    _test_ctx = ctx
+    desktop = get_desktop(ctx)
+    import uno
+    hidden_prop = uno.createUnoStruct(
+        "com.sun.star.beans.PropertyValue",
+        Name="Hidden",
+        Value=True,
+    )
+    _test_doc = desktop.loadComponentFromURL("private:factory/scalc", "_blank", 0, (hidden_prop,))
+
+@teardown
+def teardown_calc_tests(ctx):
+    global _test_doc, _test_ctx
+    if _test_doc:
+        _test_doc.close(True)
+    _test_doc = None
+    _test_ctx = None
+
+def _execute_calc_tool(name, args):
+    from plugin.main import get_tools, get_services
+    from plugin.framework.tool import ToolContext
+    # Pass suite bootstrap ctx (same as setup_calc_tests); None makes
+    # get_desktop() use uno.getComponentContext() and can segfault under
+    # python -m plugin.testing_runner.
+    tctx = ToolContext(_test_doc, _test_ctx, "calc", get_services(), "test")
+    try:
+        res = get_tools().execute(name, tctx, **args)
+    except (KeyError, ValueError) as e:
+        res = {"status": "error", "error": str(e)}
+    return res
+
+
+@native_test
+def test_charts_creation_and_listing():
+    active_sheet = _test_doc.getCurrentController().getActiveSheet()
+
+    # 1. Populate sample data
+    data = [
+        "Month", "Sales",
+        "Jan", "100",
+        "Feb", "150",
+        "Mar", "200",
+        "Apr", "250",
+        "May", "300"
+    ]
+    res_write = _execute_calc_tool("write_formula_range", {"range_name": "A1:B6", "formula_or_values": data})
+    assert res_write.get("status") == "ok", f"write_formula_range failed: {res_write}"
+
+    # 2. Create chart
+    res_create = _execute_calc_tool("create_chart", {"data_range": "A1:B6", "chart_type": "bar"})
+    assert res_create.get("status") == "ok", f"create_chart failed: {res_create}"
+
+    # 3. List charts
+    res_list = _execute_calc_tool("list_charts", {})
+    assert res_list.get("status") == "ok", f"list_charts failed: {res_list}"
+    charts = res_list.get("charts", [])
+    assert len(charts) == 1, f"Expected 1 chart, found {len(charts)}"
+    chart_name = charts[0].get("name")
+    assert chart_name is not None, "Chart name should not be None"
+
+    # 4. Query DrawPage for OLE2Shape
+    draw_page = active_sheet.getDrawPage()
+    found_chart_shape = False
+    for i in range(draw_page.getCount()):
+        shape = draw_page.getByIndex(i)
+        if shape.getShapeType() == "com.sun.star.drawing.OLE2Shape":
+            found_chart_shape = True
+            break
+    assert found_chart_shape, "com.sun.star.drawing.OLE2Shape not found on DrawPage"
+
+    # 5. Get chart info
+    res_info = _execute_calc_tool("get_chart_info", {"chart_name": chart_name})
+    assert res_info.get("status") == "ok", f"get_chart_info failed: {res_info}"
+    assert res_info.get("name") == chart_name, "Chart info name mismatch"
+
+    # 6. Edit chart
+    res_edit = _execute_calc_tool("edit_chart", {"chart_name": chart_name, "title": "Monthly Sales"})
+    assert res_edit.get("status") == "ok", f"edit_chart failed: {res_edit}"
+
+    # Verify title change
+    res_info_after_edit = _execute_calc_tool("get_chart_info", {"chart_name": chart_name})
+    assert res_info_after_edit.get("title") == "Monthly Sales", f"Chart title not updated: {res_info_after_edit}"
+
+    # 7. Delete chart
+    res_delete = _execute_calc_tool("delete_chart", {"chart_name": chart_name})
+    assert res_delete.get("status") == "ok", f"delete_chart failed: {res_delete}"
+
+    # Verify deletion
+    res_list_after_delete = _execute_calc_tool("list_charts", {})
+    assert len(res_list_after_delete.get("charts", [])) == 0, "Chart not deleted"
